fix: move `list_user_workspaces` under `v1/users`
